finished payment gateway

+++ app/Http/Livewire/QtyInCartPage.php
@@ -2,44 +2,60 @@
 
 namespace App\Http\Livewire;
 
-use App\Models\CartItem;
-use App\Models\OptionsProduct;
 use App\Repository\Cart\ModelCart;
+use App\Traits\PreviousOrder;
 use Livewire\Component;
 use App\Models\UserCart;
 
 class QtyInCartPage extends Component
 {
-    public $cart;
-    public function mount($cart)
-    {
-        $this->cart = $cart;
-    }
+    use PreviousOrder;
+    // public $cart;
+    // public function mount($cart)
+    // {
+    //     $this->cart = $cart;
+    // }
     public function render()
     {
-        $modelCart = new ModelCart();
-        $total = $modelCart->total();
-        $options = $this->cart->CartItems()->with('Options')->get();
-        return view('livewire.CartPage.qty-in-cart-page', compact('options', 'total'));
+        $cart = UserCart::first();
+        if ($cart) {
+            $modelCart = new ModelCart();
+            $total = $modelCart->total();
+            $options = $cart->CartItems()->with('Options')->get();
+            return view('livewire.CartPage.qty-in-cart-page', compact('cart', 'options', 'total'));
+        }
+        return view('livewire.CartPage.qty-in-cart-page', compact('cart'));
     }
     public function updateUp($id)
     {
+        $this->check();
         $c = new ModelCart();
         $c->updateUp(1, $id);
+        $this->emit('incermentNumber');
     }
     public function updateDown($id)
     {
+        $this->check();
         $c = new ModelCart();
         $c->updateDown(1, $id);
+        $this->emit('incermentNumber');
     }
     public function delete($id)
     {
+        $this->check();
         $c = new ModelCart();
-        $c->delete($id);
+        if (UserCart::first()->CartItems->count() == 1) {
+            $c->empty();
+        } else {
+            $c->delete($id);
+        }
+        $this->emit('incermentNumber');
     }
     public function empty()
     {
+        $this->check();
         $c = new ModelCart();
         $c->empty();
+        $this->emit('incermentNumber');
     }
 }
